add everything related to cards actions in one place
