+++ src/components/message/cards/CardActions.jsx
@@ -17,9 +17,119 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import CardTextButton from './CardTextButton';
-import CardLinkButton from './CardLinkButton';
+import { darken } from 'polished';
 
+// Actions helpers
+
+/**
+ * Gets action background color depending on component status
+ * @param props - component properties
+ * @returns {string} - the background color
+ */
+const getStatusBackgroundColor = (props) => {
+  if (props.disabled) {
+    return 'transparent';
+  }
+
+  if (props.clicked) {
+    return props.theme.colors.primary;
+  }
+
+  return 'transparent';
+};
+
+/**
+ * Gets action text color depending on component status
+ * @param props
+ * @returns {string} - the text color
+ */
+const getStatusColor = (props) => {
+  if (props.disabled) {
+    return 'lightgrey';
+  }
+
+  if (props.clicked) {
+    return props.theme.colors.primaryText;
+  }
+
+  return props.theme.colors.primary;
+};
+
+// CardLink action
+const Link = styled.a`
+  font-size: 15px;
+  font-weight: 300;
+  padding: 10px;
+  background-color: ${props => getStatusBackgroundColor(props)};
+  color: ${props => getStatusColor(props)};
+  text-align: center;
+  cursor: pointer;
+  transition: all 400ms ease;
+
+  &:hover {
+    ${props => !props.disabled && !props.clicked && `background-color: ${darken(0.03, props.theme.colors.secondary)}`};
+  }
+  &:focus {
+    outline: none;
+  }
+  ${props => (props.clicked || props.disabled ? 'pointer-events: none' : '')};
+`;
+
+const CardLink = ({ label, link, handleClick, disabled, clicked }) => (
+  <Link href={link} onClick={handleClick} target="_blank" disabled={disabled} clicked={clicked}>
+    {label}
+  </Link>
+);
+
+CardLink.propTypes = {
+  label: PropTypes.string.isRequired,
+  link: PropTypes.string.isRequired,
+  handleClick: PropTypes.func.isRequired,
+  disabled: PropTypes.bool.isRequired,
+  clicked: PropTypes.bool.isRequired,
+};
+
+// CardButton action
+const Button = styled.div`
+  font-size: 15px;
+  padding: 10px;
+  text-align: center;
+  background-color: ${props => getStatusBackgroundColor(props)};
+  color: ${props => getStatusColor(props)};
+  cursor: ${props => (props.disabled || props.clicked ? 'default' : 'pointer')};
+  transition: all 400ms ease; 
+
+  &:hover {
+    ${props => !props.disabled && !props.clicked && `background-color: ${darken(0.03, props.theme.colors.secondary)}`};
+  }
+  &:focus {
+    outline: none;
+  }
+`;
+
+const CardButton = ({ handleClick, label, disabled, clicked }) => (
+  <Button
+    onClick={disabled || clicked ? () => null : handleClick}
+    disabled={disabled}
+    clicked={clicked}
+  >
+    {label}
+  </Button>
+);
+
+CardButton.propTypes = {
+  handleClick: PropTypes.func.isRequired,
+  label: PropTypes.string.isRequired,
+  disabled: PropTypes.bool,
+  clicked: PropTypes.bool,
+};
+
+CardButton.defaultProps = {
+  disabled: false,
+  clicked: false,
+};
+
+// CardActions component
 const Container = styled.div`
   border-bottom-left-radius: 4px;
   border-bottom-right-radius: 4px;
@@ -54,7 +164,7 @@ const CardActions = ({ payload, sendAction, markAsClicked, width }) => {
         switch (action.type) {
           case 'link':
             return (
-              <CardLinkButton
+              <CardLink
                 index={index}
                 key={`${action.text}${action.linkActionValue}`}
                 link={action.linkActionValue}
@@ -67,7 +177,7 @@ const CardActions = ({ payload, sendAction, markAsClicked, width }) => {
             );
           case 'postback':
             return (
-              <CardTextButton
+              <CardButton
                 index={index}
                 key={`${action.text}${JSON.stringify(action.postbackActionValue)}`}
                 handleClick={() => {
